Skip the points for the linear fit already in the analysis to avoid lots of fit errors
Former-commit-id: 5e7256e2ee00ed3a3f13eddd7842d6127758b044

+++ Scripts/modify_config.py
@@ -20,6 +20,10 @@ elif sys.argv[1] == "RP":
     config.set('SIMULATION', 'max_mag', sys.argv[4])
     config.set('SIMULATION', 'puj_analyse_every', sys.argv[5])
     config.set('SIMULATION', 'skip_first_lf', sys.argv[6])
+    config.set('SIMULATION', 'reps_for_improvements', sys.argv[7])
+
+    if (int(sys.argv[6]) <= 5) or (int(config['SIMULATION']['skip_first_rm']) <= 5):
+        raise ValueError("At least first 5 points should be skipped since bootstrapping likely yields NaNs before")
 
     with open('config_rp.ini', 'w') as configfile:
         config.write(configfile)
